+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 -----
 
 - Adding make_public_url() function to transform a url with credential to a public one.
+- Adding set_url() function to modify a remote's url